requirements.txt
cachetools==5.1.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
google-api-core==2.8.1
google-auth==2.6.6
google-cloud-iot==2.5.0
googleapis-common-protos==1.56.2
grpc-google-iam-v1==0.12.4
grpcio==1.46.3
grpcio-status==1.46.3
idna==3.3
proto-plus==1.20.5
protobuf==3.20.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
urllib3==1.26.9
